% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_person.R
\name{get_person}
\alias{get_person}
\title{Get Person(s)}
\usage{
get_person(x, persons)
}
\arguments{
\item{x}{'character' vector.
Identifier for one or more persons.}

\item{persons}{'person' named list.
Information about an arbitrary number of persons.
Each element in the list is assigned a name, which uniquely identifies a person.}
}
\value{
A subset of \code{persons}.
}
\description{
Filter a list of individuals based on their distinct identifiers.
}
\examples{
get_person("jfisher", persons = inlpubs::authors$person)
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
